+++ src/app/inventory/DraggableInventoryItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { DragSourceSpec, DragSourceConnector, ConnectDragSource, DragSource } from 'react-dnd';
 import { DimItem } from './item-types';
-import { itemDrag, stackableDrag } from './actions';
+import { stackableDrag } from './actions';
 import store from '../store/store';
 import { BehaviorSubject } from 'rxjs';
 
@@ -33,16 +33,14 @@ let dragTimeout: number | null = null;
 const dragSpec: DragSourceSpec<Props, DragObject> = {
   beginDrag(props) {
     if (props.item.maxStackSize > 1 && props.item.amount > 1 && !props.item.uniqueStack) {
-      dragTimeout = requestAnimationFrame(() => {
-        dragTimeout = null;
-        store.dispatch(stackableDrag(true));
-      });
-    } else {
-      dragTimeout = requestAnimationFrame(() => {
-        dragTimeout = null;
-        store.dispatch(itemDrag(true));
-      });
+      store.dispatch(stackableDrag(true));
     }
+
+    dragTimeout = requestAnimationFrame(() => {
+      dragTimeout = null;
+      document.body.classList.add('drag-perf-show');
+    });
+
     isDragging = true;
     isDragging$.next(true);
     return { item: props.item };
@@ -55,9 +53,10 @@ const dragSpec: DragSourceSpec<Props, DragObject> = {
 
     if (props.item.maxStackSize > 1 && props.item.amount > 1 && !props.item.uniqueStack) {
       store.dispatch(stackableDrag(false));
-    } else {
-      store.dispatch(itemDrag(false));
     }
+
+    document.body.classList.remove('drag-perf-show');
+
     isDragging = false;
     isDragging$.next(false);
   },