+++ src/app/vendors/SingleVendor.tsx
@@ -17,7 +17,7 @@ import { refresh$ } from '../shell/refresh';
 import { InventoryBuckets } from '../inventory/inventory-buckets';
 import { connect } from 'react-redux';
 import { storesSelector, ownedItemsSelector, profileResponseSelector } from '../inventory/reducer';
-import { RootState } from '../store/reducers';
+import { RootState, ThunkDispatchProp } from '../store/reducers';
 import { toVendor } from './d2-vendors';
 import styles from './SingleVendor.m.scss';
 import vendorStyles from './Vendor.m.scss';