+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* You can now restrict what items get chosen by the random loadout feature by having an active search first. Try typing "tag:favorite" or "is:pulserifle" and then choosing Randomize.
+
 ## 5.72.0 <span className="changelog-date">(2020-03-01)</span>
 
 * Worked around a long-standing Bungie.net bug where items would change lock state when moved. One caveat is that DIM will always preserve the lock state as it sees it, so if you've locked/unlocked in game and haven't refreshed DIM, it may revert your lock.